ima: keep the integrity state of open files up to date
When a file is open for writing, kernel crash or power outage
is guaranteed to corrupt the inode integrity state leading to
file appraisal failure on the subsequent boot. Add some basic
infrastructure to keep the integrity hashes up to date as the
files are written to.

Core file operations (open, close, sync, msync, truncate)
are now allowed to update the hash immediately. In order to
maintain decent write performance, add a latency tunable
delayed work workqueue for computing the hashes.

Changelog v3:
- Squash the ima related patches
- Addressed comments from Mimi Zohar
- Isolate write tracking code into separate file(s)

Changelog v2:
- Make write measurements optional
- Add support for MMIO measurements
- Handle all writes via page flush
- Add work cancellation support, files are properly closed
  after last_writer checks out. This fixes a userspace break
  where the file was still open for writing after closing it.
- Fix/unify IMA_UPDATE_XATTR usage

Co-authored-by: Konsta Karsisto <konsta.karsisto@gmail.com>
Signed-off-by: Janne Karhunen <janne.karhunen@gmail.com>
Signed-off-by: Konsta Karsisto <konsta.karsisto@gmail.com>